pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "traitors_ai"
version = "0.1.0"
description = "LLM-driven multi-agent social deduction simulator inspired by The Traitors."
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "Your Name"}]
dependencies = [
  "langchain>=0.2.0",
  "langchain-openai>=0.1.0",
  "langchain-anthropic>=0.1.0",
  "langgraph>=0.2.0",
  "pydantic>=2.5.0",
  "typer>=0.12.0",
  "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
analysis = ["pandas>=2.0.0"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
